*: implement value transformation feature
This feature allows to add, delete and replace columns in the results
and to "modify" them.
For iperf3 that means that adding a gigabits_per_second can now be done
from ancientt itself by adding a Transformation structure such as this:
```yaml
tests:
- name: my-cool-iperf3-test
  type: iperf3
[...]
  transformations:
  - source: "bits_per_second"
    destination: "gigabits_per_second"
    action: "add"
    modifier: 100000000
    modifierAction: "division"
[...]
```

The example above causes the code to take the `bits_per_second` column
values and create a new column with the name `gigabits_per_second`.
The important part is that the `modifierAction` in this case "division"
is used with the `modifier`. So the values for the `gigabits_per_second`
column are calculated as follows: `NEW_VALUE = VALUE / MODIFIER`.

Signed-off-by: Alexander Trost <galexrt@googlemail.com>